requirements.txt
Django==1.9.7
django-modeltranslation==0.11
